[OF-1672] refac: Moves CallHelpers to common
