+++ Library/src/Multiplayer/EventBus.cpp
@@ -16,7 +16,7 @@
 #include "CSP/Multiplayer/EventBus.h"
 
 #include "CSP/Systems/SystemBase.h"
-#include "CallHelpers.h"
+#include "Common/CallHelpers.h"
 #include "Debug/Logging.h"
 #include "Multiplayer/EventSerialisation.h"
 #include "Multiplayer/SignalR/SignalRConnection.h"
