+++ Library/src/Systems/Assets/AssetSystem.cpp
@@ -19,6 +19,7 @@
 #include "CSP/Common/CSPAsyncScheduler.h"
 #include "CallHelpers.h"
 #include "Common/Algorithm.h"
+#include "Common/CallHelpers.h"
 #include "LODHelpers.h"
 #include "Multiplayer/EventSerialisation.h"
 #include "Services/PrototypeService/Api.h"
